moves functions to more appropriate module
